update: fixing delete
