+++ src/context/features/articleManagementSlice.ts
@@ -1,10 +1,8 @@
 import { createSlice } from "@reduxjs/toolkit";
 import { PayloadAction } from "@reduxjs/toolkit";
-import { ArticleData, CategoryData } from "../../types/articleManagement.type";
+import { ArticleData } from "../../types/articleManagement.type";
 
 const initialState  = {
-  categoryData: { pages: {}, searchResults: {}},
-  categoryDataMeta: { next_page_url: '', prev_page_url: '', total: 0, per_page: 0, current_page: 0, last_page: 0 },
   articleData: { pages: {}, searchResults: {}},
   articleDataMeta: { next_page_url: '', prev_page_url: '', total: 0, per_page: 0, current_page: 0, last_page: 0 }
 };
@@ -13,23 +11,6 @@ const articleManagementSlice = createSlice({
   name: 'articleManagement',
   initialState,
   reducers: {
-    setCategoryData(state: { categoryData: { pages: { [key: string]:  CategoryData[]}; searchResults: {[key:string]: {[key:string]:  CategoryData[]}}}}, action: PayloadAction<{ page: string; data: CategoryData[]; search: string | undefined }>) {
-      const { page, data, search } = action.payload;
-
-      if (search) {
-        if (!state.categoryData.searchResults[search]) {
-          state.categoryData.searchResults[search] = {};
-        }
-        state.categoryData.searchResults[search][page] = data;
-      } else {
-        state.categoryData.pages[page] = data;
-      }
-    },
-
-    setCategoryDataMeta: (state, action) => {
-        state.categoryDataMeta = action.payload;
-    },
-
     setArticleData(state: { articleData: { pages: { [key: string]:  ArticleData[]}; searchResults: {[key:string]: {[key:string]:  ArticleData[]}}}}, action: PayloadAction<{ page: string; data: ArticleData[]; search: string | undefined }>): void {
       const { page, data, search } = action.payload;
 
@@ -48,13 +29,11 @@ const articleManagementSlice = createSlice({
     },
 
     clearArticleManagementData: (state) => {
-      state.categoryData = { pages: {}, searchResults: {}};
-      state.categoryDataMeta = { next_page_url: '', prev_page_url: '', total: 0, per_page: 0, current_page: 0, last_page: 0 };
       state.articleData = { pages: {}, searchResults: {}};
       state.articleDataMeta = { next_page_url: '', prev_page_url: '', total: 0, per_page: 0, current_page: 0, last_page: 0 };
     },
   },
 });
 
-export const { setCategoryData, setCategoryDataMeta, setArticleData, setArticleDataMeta, clearArticleManagementData } = articleManagementSlice.actions;
+export const { setArticleData, setArticleDataMeta, clearArticleManagementData } = articleManagementSlice.actions;
 export default articleManagementSlice.reducer;
