+++ src/hooks/articleManagement/useFetchCategoryDelete.ts
@@ -2,8 +2,9 @@ import { useEffect, useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 import { setLoading } from '../../context/features/appSlice';
 import HandleRespondResponse from '../../services/utils/handleRespondResponse';
-import { CategoryData } from '../../types/articleManagement.type';
 import { deleteCategory } from '../../api/articleManagement';
+import { CategoryData } from '../../types/categoryManagement.type';
+import { clearCategoryManagementData } from '../../context/features/categoryManagementSlice';
 
 const useFetchCategoryDelete = (id: number | undefined, onSuccess?: () => void) => {
   const dispatch = useDispatch();
@@ -14,8 +15,9 @@ const useFetchCategoryDelete = (id: number | undefined, onSuccess?: () => void) 
       dispatch(setLoading(true));
       const response = await deleteCategory(id)
       dispatch(setLoading(false));
-      HandleRespondResponse<CategoryData>(response, 'deleted');
-      if (onSuccess) {
+      HandleRespondResponse<CategoryData>(response, 'deleted');      
+      dispatch(clearCategoryManagementData());
+      if (onSuccess) {        
         onSuccess();
       }
     } catch (error) {
