+++ src/types/app.type.ts
@@ -1,5 +1,6 @@
-import { ArticleData, CategoryData } from "./articleManagement.type";
+import { ArticleData } from "./articleManagement.type";
 import { UserStateContext } from "./auth.type";
+import { CategoryData } from "./categoryManagement.type";
 import { PermissionData } from "./userManagement.type";
 import { RoleData } from "./userManagement.type";
 import { UserData } from "./userManagement.type";
@@ -30,17 +31,21 @@ export type userManagementStateContext = {
 }
 
 export type articleManagementStateContext = {
-    categoryData: {
-        pages: { [key: string]: CategoryData[] },
-        searchResults: { [key: string]: { [key: string]: CategoryData[] } }
-    }; 
-    categoryDataMeta: MetaType;
     articleData:  {
         pages: { [key: string]: ArticleData[] },
         searchResults: { [key: string]: { [key: string]: ArticleData[] } }
     };
     articleDataMeta: MetaType;
 }
+
+export type categoryManagementStateContext = {
+    categoryData: {
+        pages: { [key: string]: CategoryData[] },
+        searchResults: { [key: string]: { [key: string]: CategoryData[] } }
+    }; 
+    categoryDataMeta: MetaType;
+}
+
 
 export type AppContext = {
     isLoading: boolean,
@@ -55,5 +60,6 @@ export type StateContext = {
     user: UserStateContext;
     userManagement: userManagementStateContext;
     articleManagement: articleManagementStateContext;
+    categoryManagement: categoryManagementStateContext;
     app: AppContext;
 }
