+++ src/types/categoryManagement.type.ts
@@ -1,0 +1,7 @@
+export type CategoryData =  {
+    id: number,
+    name: string,
+    description: string,
+    created_at: string,
+    updated_at: string
+}
